fix: Dropdown generics
Create a workaround for keeping generic value type of Dropdown component when creating a React wrapper.

Note: this React component is unusable until this bug is fixed in the web component repository https://github.com/fintraffic-design/fds-coreui-components/issues/65 but this change has been tested with a locally patched version of the web component.